refactor(c2pa): move merkle map extraction to the merkle module
