+++ libs/c2pa/src/merkle/MerkleSegmentValidation.ts
@@ -2,20 +2,16 @@ import type { C2paStatusCode } from '../C2paStatusCode.ts'
 import type { LiveVideoStatusCode } from '../LiveVideoStatusCode.ts'
 
 /**
- * One entry from the `merkle` array of the `c2pa.hash.bmff.v3` assertion (§18.6).
+ * One merkle map from the `c2pa.hash.bmff.v3` assertion (§18.6).
  *
  * @public
  */
 export type MerkleMap = {
 	readonly uniqueId: number
 	readonly localId: number
-	/** Number of leaf nodes (media segments) in the tree */
 	readonly count: number
-	/** Merkle tree row stored in the manifest */
 	readonly hashes: readonly Uint8Array[]
-	/** `null` for single-file fMP4 */
 	readonly initHash: Uint8Array | null
-	/** `null` for the default (SHA-256) */
 	readonly alg: string | null
 	readonly exclusions: readonly {
 		readonly xpath: string
@@ -24,7 +20,6 @@ export type MerkleMap = {
 			readonly value: Uint8Array | readonly number[]
 		}[]
 	}[]
-	/** `8`, or `0` when the assertion also carries `hash` (§18.6.2) */
 	readonly offsetPrefixSize: number
 }
 
@@ -35,7 +30,6 @@ export type MerkleMap = {
  * @public
  */
 export type MerkleSegmentState = {
-	/** Last validated location per `${uniqueId}:${localId}` track key */
 	readonly lastLocation: ReadonlyMap<string, number>
 }
 
@@ -45,9 +39,7 @@ export type MerkleSegmentState = {
  * @public
  */
 export type MerkleSegmentValidation = {
-	/** Zero-based leaf index, or `null` if malformed */
 	readonly location: number | null
-	/** Computed leaf hash, hex-encoded */
 	readonly bmffHashHex: string | null
 	readonly isValid: boolean
 	readonly errorCodes: readonly (LiveVideoStatusCode | C2paStatusCode)[]
